refactor(shells): modularize shells
The first version of this setup had a number of dependencies that
probably didn't generalize to other setups. This makes it so most of the
shell based dependencies can drop a file in something like `~/.zshrc.d/`
or `~/.config/fish/fish.d` and those files will be picked up by the
shell. This results in a lot more flexible of a shell setup.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>