feat(powerline): add fish duration patch
fish has built in support for the duration of commands, but it only goes
down to the millisecond level and not microsecond. This creates an
optional patch that avoids sending 0μs on every command. It can
activated by setting `powerline_go.fish_duration_patch` to true.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>