docs(powerline): add README file
This is a pretty basic file for the powerline-go role. There's a few
other spots where there are variables that can be defined, but it
covers the main ones.

I probably need to find a way to connect it to the roles for `fish` and
`zsh` that I've created as the shell configuration only works with those
roles.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>