feat(fish): add support for custom configuration
This adds support for the custom key `fish_additional` which provides a
small spot for additional code to be dropped in the powerline-go
configuration for fish. This is designed to be used to correct syntax
highlighting issues in powerline-go which can cause some servers to look
like they're showing errors in the prompt.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>