fix(fish): update duration patch for fish
In most cases you don't need to know execution time down to the
millisecond and fish doesn't track it that far - so it would always show
0ns for everything. That's annoying. This patch reduces precision of
time to eliminate that annoyance in fish.

In reality, it's just a periodic update of the patch.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>