fix(fish): correct shell evaluation for fish
There were a couple of different problems with the shell evaluation for
fish. Notably, you don't need the `-shell bare` anymore. Which is
awesome! Next, I noticed that I wasn't handling the error codes properly
because there's statement in the prompt that invokes an executable and
caused problems with the status segment. This fixes both of those.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>