fix(fish): correct duration and errors
I wanted Fish and Powerline-go to work such that the duration and error
codes are only displayed the first time you run the command. This has
the side effect of nuking the `$CMD_DURATION` environment var and
replacing it with `$LAST_CMD_DURATION`, but it also makes it so the
error on your prompt only shows up the first time you see the error.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>