fix(pwsh): fix powershell directory information
`pwsh_shell_script_path` is a directory because it may need to drop in
multiple files - both a default profile and the powerline-go prompt
code. However, it was still getting `dirname` called on it, which would
cause it to fail if the directory was not yet present.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>